Se empezo con la modificacion referido a la carga de archivos, ademas se cambio el trabajo a uno MAVEN para poder trabajar con JSON y GSON
