+++ ProyectoJuegoJava/ProjectoFinal/Final/src/main/java/Trabajo/Personajes/Mago.java
