--- ProyectoJuegoJava/ProjectoFinal/src/Trabajo/Enemigos/Enemigo.java
@@ -1,6 +1,0 @@
-package Trabajo.Enemigos;
-
-public class Enemigo {
-    private String nombre;
-    private
-}
